dev: extract secure
